requirements.txt
streamlit
pandas
pyarrow
Pillow